ml/requirements.txt
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
langgraph==0.2.14
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
duckduckgo-search==6.2.13
beautifulsoup4==4.12.3
requests==2.32.3
reportlab==4.2.2
python-dotenv==1.0.1
pydantic==2.8.2
arxiv==2.1.3
httpx==0.27.0